:memo: docs: Add Deployment Architecture (part of #192)
This commit adds a new documentation page with our Deployment
Architecture into the Admin Documentation chapter of our full
documentation. The point of this page is to communicate our expected
vision of how this application should be managed and deployed. Since
Monroe County has not previously worked with containers but is
interested in doing so, this tries to help give them a secure and simple
pathway of proving this technology in their own infrastructure.

One-third of #192. This commit completes my deliverables for that issue.

Signed-off-by: Justin W. Flory <git@jwf.io>