:memo: docs: Refactor "Features" page to "Features and Systems"
Part of #123, even though it's documentation.

This commit is a final refactor of our Features page. I updated statuses
on all requirements, reworded the page introduction text at the top, and
removed technology decisions that were outdated to our final
implementation. This document is accurate to-date and I do not
anticipate needing to change it again before hand-off.

Signed-off-by: Justin W. Flory <git@jwf.io>