:bowtie: Reformat all Python code to Black standard (closes #123)
This commit refactors all of the remaining Python code to follow the
standards set by the Python Software Foundation's "Black" tool:

https://github.com/psf/black

This are a lot of changes, but all of the changes are supposed to be
cosmetic. These changes change the appearance of the Python code to
adhere to the PEP-8 standards (with some deviations):

https://pep8.org/

I used the `code_lint.sh` script to automatically clean up all the code.
To better automate this in the future, I make a new pull request with a
git pre-commit hook so Black is run every time before a commit is made.
This way, we don't have to worry about following the standard
consistently; it will be mandated for us.

Closes #123. Huzzah! :tada:

Signed-off-by: Justin W. Flory <git@jwf.io>