Update helpers.py
This covers the helpers.py file comments.

In addition:
helpers.py: Changed the change_username function. Originally it was using the request to get the old email address and request.POST to get the new email address. This is not nessasary, the two emails are already arguments. The function now grabs from that argument.