:unicorn: containers: Change entrypoint to use Gunicorn (closes #250)
This commit changes the entrypoints on the Dockerfiles to use `gunicorn`
instead of the Django debug server. I was able to switch it for both
development and production since Gunicorn is fairly simple.

For development, I included a setting that automatically reloads
Gunicorn if files change on disk. This is a setting intended for local
development, such as when we have container volumes with docker-compose.

Also, I have waited a long time to have a semi-valid excuse to use the
:unicorn: emoji.

Closes #250.

Signed-off-by: Justin W. Flory <git@jwf.io>