:memo: docs: Refactor Admin Guide into ReStructuredText
This is a pretty drastic rewrite of the quickstart guide into
ReStructuredText. In addition to converting the format and adding some
of the fancier text formatting things, I also refactored the document to
be more direct in a few places.

I also added an explainer about Podman in our dependencies section.

Signed-off-by: Justin W. Flory <git@jwf.io>